+++ mgi-pulse-core/src/engine/mod.rs
@@ -20,6 +20,7 @@ pub mod parse;
 pub mod parse_edn;
 pub mod parse_logfmt;
 pub mod parse_python;
+pub mod parse_syslog;
 pub mod predicate;
 pub mod query;